refactor(uffd-tests): move internal/rpcharness → testutils/testharness
Sibling of the existing testutils helpers; rename + move in one step.
The path now makes the test-only origin obvious. Drops internal/
scoping but no production caller exists or would want this package.
