+++ packages/orchestrator/pkg/sandbox/uffd/testutils/testharness/barriers.go
@@ -1,4 +1,4 @@
-package rpcharness
+package testharness
 
 import (
 	"context"
